requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM APIs - OpenAI required for embeddings
openai>=1.3.0

# Optional LLM providers (choose at least one)
groq>=0.30.0
# google-generativeai>=0.8.0  # Uncomment if using Gemini

# Vector Store
chromadb>=0.4.22

# Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.26.0

# Progress bars
tqdm>=4.65.0